--- old-files/Pipfile
@@ -1,17 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-opencv-python = "==4.7.0.68"
-mediapipe = ">=0.9.0.1"
-scikit-learn = "==1.2.0"
-pymongo = "*"
-requests = "*"
-flask = "*"
-
-[dev-packages]
-
-[requires]
-python_version = "3.10"
