+++ web-app/Dockerfile
@@ -2,6 +2,11 @@ FROM python:3.8-slim
 
 WORKDIR /app
 
+RUN apt-get update && apt-get install -y \
+    libgl1-mesa-glx \
+    libglib2.0-0 \
+    && rm -rf /var/lib/apt/lists/*
+
 # Install dependencies
 COPY requirements.txt .
 RUN pip install --no-cache-dir -r requirements.txt
