+++ web-app/ReadMe.md
@@ -1,0 +1,2 @@
+NEED TO RUN THIS ONCE RUNNING CONTAINERS TO ACCESS THE WEB APP: 
+http://localhost:5003
