+++ web-app/requirements.txt
@@ -1,3 +1,6 @@
 flask==2.0.1
 werkzeug==2.0.2
-requests==2.26.0
+requests==2.26.0 
+numpy>=1.19.0 
+opencv-python==4.7.0.68  
+pillow>=8.0.0
