+++ web-app/templates/index.html
@@ -1,73 +1,91 @@
 <!DOCTYPE html>
-<html>
+<html lang="en">
 <head>
-    <title>MacOS Webcam Stream</title>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>ASL Recognition</title>
     <style>
         body {
             font-family: Arial, sans-serif;
             margin: 0;
             padding: 20px;
             text-align: center;
-            background-color: #f5f5f5;
-        }
-        h1 {
-            color: #333;
         }
         .container {
             max-width: 800px;
             margin: 0 auto;
         }
-        .video-container {
-            margin-top: 20px;
-            border: 1px solid #ddd;
-            padding: 10px;
-            border-radius: 5px;
-            background-color: white;
-            box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+        #videoContainer {
+            position: relative;
+            margin: 20px auto;
+            width: 640px;
+            height: 480px;
+            border: 1px solid #ccc;
         }
-        .stream-img {
+        #video {
             width: 100%;
-            max-width: 640px;
-            height: auto;
+            height: 100%;
+            background-color: #f0f0f0;
         }
-        .info {
+        #processedVideo {
+            width: 100%;
+            height: 100%;
+            background-color: #f0f0f0;
+            display: none;
+        }
+        #recognition {
+            font-size: 2em;
+            margin: 20px 0;
+            height: 1.5em;
+        }
+        .controls {
+            margin: 20px 0;
+        }
+        button {
+            padding: 10px 20px;
+            margin: 0 10px;
+            cursor: pointer;
+            background-color: #4CAF50;
+            color: white;
+            border: none;
+            border-radius: 4px;
+        }
+        button:hover {
+            background-color: #45a049;
+        }
+        button:disabled {
+            background-color: #cccccc;
+            cursor: not-allowed;
+        }
+        .status {
             margin-top: 20px;
-            padding: 10px;
-            background-color: #e9f7fe;
-            border-radius: 5px;
-            text-align: left;
-        }
-        .loading {
-            padding: 40px;
-            font-style: italic;
-            color: #666;
+            color: #555;
         }
     </style>
 </head>
 <body>
     <div class="container">
-        <h1>MacOS Webcam Stream</h1>
-        <div class="video-container">
-            <div class="loading">Loading webcam stream...</div>  
-            <img class="stream-img" src="/stream" alt="Webcam Stream">
-        </div>
-
-
+        <h1>ASL Sign Language Recognition</h1>
         
-        <div class="info">
-            <h3>How this works:</h3>
-            <p>1. A Python script on your Mac is accessing your webcam</p>
-            <p>2. It streams the webcam feed over HTTP on port 5002</p>
-            <p>3. This Docker container connects back to your Mac to get the stream</p>
-            <p>4. The container serves this webpage which displays the stream</p>
+        <div id="videoContainer">
+            <video id="video" autoplay playsinline></video>
+            <img id="processedVideo" alt="Processed video">
+        </div>
+        
+        <div id="recognition">
+            <!-- Recognized sign will be displayed here -->
+        </div>
+        
+        <div class="controls">
+            <button id="startBtn">Start Camera</button>
+            <button id="stopBtn" disabled>Stop Camera</button>
+        </div>
+        
+        <div class="status" id="status">
+            Camera not started
         </div>
     </div>
 
-    <script>
-        // Hide the loading message when the image loads
-        document.querySelector('.stream-img').onload = function() {
-            document.querySelector('.loading').style.display = 'none';
-        };
-    </script>
+    <script src="/static/app.js"></script>
 </body>
 </html>
