--- working on/process.txt
@@ -1,12 +1,0 @@
-1. need to use version 3.8.20 of pipenv (CHECK PIPENV VERSION USING pipenv --version) when installing the pipfile in the root file. I used pyenv 3.8 to create a temporary shell has 3.8 downloaded to then create a pipenv specifically using python version 3.8. This ensures that the proper python version is used to install the dependencies. 
-
-2. ensure your pipenv is running in python 3.8, the version that matches the python version in the docker file. Use pyenv to install version 3.8, then create a pyenv shell: pyenv shell 3.8.20
-
-3. inside the shell check the python version python --version. it should say 3.8.20 
-
-4. run: pipenv --python $(pyenv which python)  
-  - this installs pipenv using the specified version of python 3.8.20 
-
-5. run pipfile install to now create a pipenv that is in python 3.8 (the version in the pipfile) 
-
-6. Verify the venv was created in python 3.8 by using command+shift+p
